Began Easy mode, various UI fixes.
Fixed capitalise and regionise being executed in the wrong order, resulting in hyphens in the name of delta pokemon. Also added a temporary css file at \src\routes\(app)\ and imported it into +page.svelte to load images for the easy mode of insurgence. as I can't change the insurgence.css file on the server to insurgence_e.css. Added data for mega insurgence pokemon to patches.json. Tbh, not sure if this was entirely necessary still don't have a great understanding of the data structures at work.